+++ src/solvers/QP/QPSolverFactory.hpp
@@ -7,7 +7,8 @@
 
 class QPSolverFactory {
 	public:
-		static std::unique_ptr<QPSolver> create(const std::string& QP_solver_name, int number_variables, int number_constraints, int maximum_number_nonzeros, bool quadratic_programming);
+		static std::unique_ptr<QPSolver> create(const std::string& QP_solver_name, size_t number_variables, size_t number_constraints,
+		      size_t maximum_number_nonzeros, bool quadratic_programming);
 };
 
 #endif // QPSOLVERFACTORY_H
